Address review comments - 2
