+++ ql/src/java/org/apache/hadoop/hive/ql/plan/FetchWork.java
@@ -84,11 +84,6 @@ public class FetchWork implements Serializable {
 
   private Set<FileStatus> filesToFetch = null;
 
-  /**
-   * Whether this FetchWork should enable cache
-   */
-  private boolean cachingEnabled = false;
-
   public boolean isHiveServerQuery() {
 	return isHiveServerQuery;
   }
@@ -401,13 +396,5 @@ public class FetchWork implements Serializable {
 
   public Set<FileStatus> getFilesToFetch() {
     return filesToFetch;
-  }
-
-  public boolean isCachingEnabled() {
-    return cachingEnabled;
-  }
-
-  public void setCachingEnabled(boolean cachingEnabled) {
-    this.cachingEnabled = cachingEnabled;
   }
 }
